refactor(ReservationRepository): 인기 테마 조회 DTO 추가
