+++ src/main/java/roomescape/reservation/repository/PopularThemeQueryResult.java
@@ -1,0 +1,4 @@
+package roomescape.reservation.repository;
+
+public record PopularThemeQueryResult(String name, String description, String thumbnailUrl) {
+}
