+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -1,9 +1,9 @@
 package roomescape.reservation.repository;
 
+import roomescape.reservation.domain.Reservation;
+
 import java.time.LocalDate;
 import java.util.List;
-import roomescape.reservation.domain.Reservation;
-import roomescape.theme.domain.Theme;
 
 public interface ReservationRepository {
 
@@ -15,5 +15,5 @@ public interface ReservationRepository {
 
     boolean existsByDateAndTimeIdAndThemeId(LocalDate date, Long timeId, Long themeId);
 
-    List<Theme> findPopularThemes(int period, int limit);
+    List<PopularThemeQueryResult> findPopularThemes(int period, int limit);
 }
