+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,14 +1,17 @@
 package roomescape.reservation.service;
 
-import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
+import roomescape.reservation.service.dto.PopularThemesResult;
+import roomescape.reservation.service.dto.ReservationCommand;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.repository.ThemeRepository;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.repository.ReservationTimeRepository;
+
+import java.util.List;
 
 @Service
 @Transactional(readOnly = true)
@@ -52,8 +55,7 @@ public class ReservationService {
         return reservationRepository.findAll();
     }
 
-    public List<Theme> findPopularThemes(int period, int limit) {
-        return reservationRepository.findPopularThemes(period, limit);
-
+    public PopularThemesResult findPopularThemes(int period, int limit) {
+        return new PopularThemesResult(reservationRepository.findPopularThemes(period, limit));
     }
 }
