+++ src/main/java/roomescape/reservation/service/dto/PopularThemesResult.java
@@ -1,0 +1,8 @@
+package roomescape.reservation.service.dto;
+
+import roomescape.reservation.repository.PopularThemeQueryResult;
+
+import java.util.List;
+
+public record PopularThemesResult(List<PopularThemeQueryResult> popularThemes) {
+}
