+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -1,22 +1,16 @@
 package roomescape.theme.controller;
 
-import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import roomescape.reservation.service.ReservationService;
 import roomescape.theme.controller.dto.ThemeRequest;
 import roomescape.theme.controller.dto.ThemeResponse;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.service.ThemeService;
+
+import java.net.URI;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/themes")
@@ -44,7 +38,7 @@ public class ThemeController {
             @RequestParam("period") int period,
             @RequestParam("limit") int limit
     ) {
-        List<ThemeResponse> responses = reservationService.findPopularThemes(period, limit)
+        List<ThemeResponse> responses = reservationService.findPopularThemes(period, limit).popularThemes()
                 .stream()
                 .map(ThemeResponse::from)
                 .collect(Collectors.toList());
@@ -57,7 +51,7 @@ public class ThemeController {
         Theme theme = themeService.save(requestDto.toCommand());
         ThemeResponse response = ThemeResponse.from(theme);
         return ResponseEntity
-                .created(URI.create("/themes/" + response.id()))
+                .created(URI.create("/themes/" + theme.getId()))
                 .body(response);
     }
