+++ src/main/java/roomescape/theme/controller/dto/ThemeRequest.java
@@ -1,6 +1,6 @@
 package roomescape.theme.controller.dto;
 
-import roomescape.theme.service.ThemeCommand;
+import roomescape.theme.service.dto.ThemeCommand;
 
 public record ThemeRequest(String name, String description, String thumbnailUrl) {
