+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.repository.ThemeRepository;
+import roomescape.theme.service.dto.ThemeCommand;
 import roomescape.time.domain.ReservationTime;
 
 import java.time.LocalDate;
