+++ src/main/java/roomescape/theme/service/dto/ThemeCommand.java
@@ -1,4 +1,4 @@
-package roomescape.theme.service;
+package roomescape.theme.service.dto;
 
 public record ThemeCommand(String name, String description, String thumbnailUrl) {
 }
